added downloading company by id

+++ provider/src/main/java/com/tkolbusz/provider/dto/converter/MoneyConverter.java
@@ -4,15 +4,12 @@ import com.tkolbusz.domain.model.Currency;
 import com.tkolbusz.domain.model.Money;
 import com.tkolbusz.domain.util.BasicConverter;
 
-public class MoneyConverter extends BasicConverter<Money, Integer> {
+public class MoneyConverter extends BasicConverter<Money, Double> {
     @Override
-    public Money transform(Integer DTO) {
+    public Money transform(Double DTO) {
         if (DTO == null) {
             return null;
         }
-        return new Money(
-                DTO,
-                Currency.PLN
-        );
+        return new Money(DTO, Currency.PLN);
     }
 }
